generator: Add c_function (C function name) to each action struct.
This field, which is generated internally by the generator, is the
name of the C function corresponding to each action.

For actions that have NO optional arguments, it's just "guestfs_<name>".

For actions that have any optional arguments, it is
"guestfs_<name>_argv" (since any binding has to construct the optional
argument struct explicitly).

In a future commit, this mapping may become more complex.

This commit also "fixes" the C# bindings which didn't handle optional
arguments properly at all.  In fact, it doesn't fix this, it just
changes it enough that it probably now compiles.  We should either
compile and test the bindings routinely with Mono, or drop them, since
they are starting to bit-rot.

In the GObject bindings, I have added a space between the C function
name and the first paren.

Apart from the C# and GObject changes, this is just code motion.  It
was verified by diffing the output of the generator before and after.