run: Label tmp directory.
There's not really any point just documenting that people
should label the tmp directory, when we can just label it.
So change the ./run script so it calls chcon.

However it's not a failure if chcon doesn't work.  If SELinux
is not installed or is disabled, then labelling will fail.