proto: Free drives array when qemu subprocess dies abruptly.
Commit ed7fda161e1f3d0beb02a368fcbcf5ed95dcdac1 changed the g->drives
linked list into an array, and also made a subtle change to the
(undocumented) semantics of drives in the rare case when qemu is
relaunched.

Every time we go back to the CONFIG state, we should free the drives
array.  However there was one place where this was not happening.
This was caught by the tests of the protocol, but only when using the
default (appliance) method, not the libvirt method which we were more
interested in testing in Fedora.

When testing on Debian using the appliance method, we therefore hit
this bug for the first time.

This relates to or fixes the following commits:
  commit 9df159b34a9e41e6f36c5435869a54396b92a864
  commit 33f49d85c2a82e66f33cedccb2dbd14faa6070dd
  commit ed7fda161e1f3d0beb02a368fcbcf5ed95dcdac1