launch: appliance: Fix parsing of QEMU_OPTIONS.
If there was no quoting character, the options could run off the end
of the string and into whatever string which happened to be following
in memory.

This bug was revealed when libguestfs was compiled on arm.
(cherry picked from commit 6c9aaf394f078b924a46a5b2e4ab1b3576d17b72)