proto: Fix for rare FileIn hangs. (RHBZ#969845).
Yet another protocol bug!

If the daemon cancels, then the library should send a cancellation
chunk.  It normally does this during the loop where it is uploading a
file.  However due to an oversight or possibly a regression when we
modularized the socket code, the library did not send a cancellation
chunk if the daemon cancelled just before the end of file was sent.

This commit fixes this.
(cherry picked from commit 697b0f89fed04a4017e033861a9eaaa40b3270d9)