golang: Fix it so it builds if libguestfs is not installed.
It was using the installed copy of libguestfs not the local copy.

We can't encode the paths in guestfs.go, since this source file gets
copied into the install directory where those paths would not make
sense.  Instead we have to use CGO_CFLAGS and CGO_LDFLAGS to
communicate the right flags to go when doing the local build and test.