info: Fix double-call to error function when qemu-img info command fails (RHBZ#981663).
Previously if the 'qemu-img info --output json' command failed, then
the stdout callback would be called first with a zero-length buffer.
Since a zero-length buffer cannot be valid JSON (in this context)
don't produce a separate error message.

However do catch the case when qemu-img info might produce no output
and no error (if that were to happen -- it does not happen at the
moment).