Fix --enable-valgrind-daemon option.
Don't add the "valgrind channel" to the appliance.

Just dump out the valgrind.log to stderr while the daemon is running.

Ensure that if valgrind tests fail in the appliance, that we don't
exit with success in the library by checking for a canary message in
the verbose daemon logs.

This allows the option to be used routinely by developers.

(cherry picked from commit 55e3b8711f340a2f8bdb8ee8ff99deb40b4e9108)