launch: direct: Don't try to wait for qemu if parent process forked (RHBZ#998482).
When using guestfish --remote, libguestfs is not the parent of qemu,
so waitpid is expected to return -ECHILD.  And indeed guestfish --remote
would print a bogus error message in this case:

  libguestfs: error: waitpid (qemu): No child processes

If the parent process forked into the background, then it should have
called 'set_recovery_proc 0' so we can use this to determine if we
need to wait for qemu.

Thanks: Kazuya Saito for reporting and identifying the issue.

(cherry picked from commit a843b5e5e32c151e7b0c74bb4f7be1030f9ac85b)